refactor(My Employees): Flyttet filer over til komponenter og byttet til ServerSideProps
